agregar configuracion de eslint y sistema de gestion de roles en paginas

+++ graphql/models/client/resolvers.ts
@@ -1,13 +1,22 @@
 import prisma from 'config/prisma';
 
 const ClientResolvers = {
+  Client: {
+    createdAt: async (parent, _, context) => {
+      if (context.session.user.role.name === 'Admin') {
+        return parent.createdAt;
+      }
+
+      return null;
+    },
+  },
   Query: {
-    getClients: async (parent, args) => {
+    getClients: async () => {
       return await prisma.client.findMany();
     },
   },
   Mutation: {
-    createClient: async (parent, args) => {
+    createClient: async (_, args) => {
       const nuevoCliente = await prisma.client.create({
         data: {
           name: args.name,
@@ -15,7 +24,7 @@ const ClientResolvers = {
       });
       return nuevoCliente;
     },
-    updateClient: async (parent, args) => {
+    updateClient: async (_, args) => {
       return await prisma.client.update({
         where: {
           id: args.id,
@@ -27,12 +36,15 @@ const ClientResolvers = {
         },
       });
     },
-    deleteClient: async (parent, args) => {
-      return await prisma.client.delete({
-        where: {
-          id: args.id,
-        },
-      });
+    deleteClient: async (_, args, context) => {
+      if (context.session.user.role.name === 'Admin') {
+        return await prisma.client.delete({
+          where: {
+            id: args.id,
+          },
+        });
+      }
+      return null;
     },
   },
 };
